Unify testing of atomic db implementations
